To be DRY: abstract DAO class Untested changes, may break the code! Kaizen.db changed?
Will PR now so Marcin can take a look and offer pointers in fixing.

First, solve the problem. Then, write the code.
      -- John Johnson
